geektime_0-master/requirements.txt
allure-pytest==2.12.0
allure-python-commons==2.11.0
Appium-Python-Client==2.7.0
asgiref==3.5.2
async-generator==1.10
attrs==22.1.0
blinker==1.4
Brotli==1.0.9
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
cryptography==38.0.1
exceptiongroup==1.0.0rc9
Flask==2.2.2
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.0
gitdb==4.0.9
GitPython==3.1.29
greenlet==1.1.3
h11==0.13.0
h2==4.1.0
hpack==4.0.0
hyperframe==6.0.1
idna==3.4
iniconfig==1.1.1
itsdangerous==2.1.2
jenkinsapi==0.3.11
Jinja2==3.1.2
jsonpath==0.82
kaitaistruct==0.10
ldap3==2.9.1
MarkupSafe==2.1.1
mitmproxy==8.0.0
mitmproxy_wireguard==0.1.14
msgpack==1.0.4
outcome==1.2.0
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
protobuf==4.21.7
publicsuffix2==2.20191221
py==1.11.0
pyasn1==0.4.8
pycparser==2.21
pydantic==1.10.2
PyJWT==2.5.0
PyMySQL==1.0.2
pyOpenSSL==22.1.0
pyparsing==3.0.9
pyperclip==1.8.2
PySocks==1.7.1
pytest==7.1.3
pytz==2022.4
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.5.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.41
tomli==2.0.1
tornado==6.2
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.12
urwid==2.1.2
Werkzeug==2.2.2
wsproto==1.1.0
zstandard==0.18.0
allure-python-commons==2.12.0
